Adicionando tratamento de erros

+++ src/Controllers/PessoaController.cs
@@ -1,7 +1,9 @@
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
+using System.Net;
+
 using src.Models;
 using src.Persistence;
-using Microsoft.EntityFrameworkCore;
 
 namespace src.Controllers;
 
@@ -16,38 +18,75 @@ public class PessoaController : ControllerBase{
     }
 
     [HttpGet]
-    public List<Pessoa> get(){
-        //Pessoa pessoa = new Pessoa("ale", 23, "12345678");
-        //Contrato contrato = new Contrato("abc123", 50.43);
-        //pessoa.Contratos.Add(contrato);
+    public ActionResult<List<Pessoa>> Get(){
+        var result = _context.Pessoas.Include( p => p.Contratos).ToList();
 
-        return _context.Pessoas.Include( p => p.Contratos).ToList();
+        if(!result.Any())
+            return NoContent();
+        return Ok(result);
     }
 
     [HttpPost]
-    public Pessoa Post([FromBody]Pessoa pessoa){
-        _context.Pessoas.Add(pessoa);
-        _context.SaveChanges();
+    public ActionResult<Pessoa> Post([FromBody] Pessoa pessoa){
+        
+        try{
+            _context.Pessoas.Add(pessoa);
+            _context.SaveChanges();
+        
+        } catch(SystemException){
+            return BadRequest(new {
+                msg = "Houve um erro ao tentar criar!",
+                status = HttpStatusCode.BadRequest
+            });
+        }
 
-        return pessoa;
+        return Created("Cadastro criado!", pessoa);
     }
 
     [HttpPut("{id}")]
-    public string Update([FromRoute]int id, [FromBody]Pessoa pessoa){
-        _context.Pessoas.Update(pessoa);
-        _context.SaveChanges();
+    public ActionResult<Object> Update([FromRoute] int id, [FromBody] Pessoa pessoa){
+        var result = _context.Pessoas.SingleOrDefault(e => e.Id == id);
 
-        return "Dados do id " + id + " atualizados!";
+        if(result is null)
+            return NotFound(new {
+                msg = "Registro não encontrado",
+                status = HttpStatusCode.NotFound
+            });
+
+        try{
+            _context.Pessoas.Update(pessoa);
+            _context.SaveChanges();
+        
+        } catch(SystemException){
+            return BadRequest( new{
+                msg = "Houve erro ao tentar atualizar o id " + id,
+                status = HttpStatusCode.BadRequest
+            });
+        }
+        
+        return Ok(new {
+            msg = "Dados do id " + id + " atualizados!",
+            status = HttpStatusCode.OK
+        });
     }
 
     [HttpDelete("{id}")]
-    public string Delete([FromRoute]int id){
+    public ActionResult<Object> Delete([FromRoute] int id){
         var result = _context.Pessoas.SingleOrDefault(e => e.Id == id);
 
+        if(result is null)
+            return BadRequest(new {
+                msg = "Conteúdo inexistente, solicitação inválida!",
+                status = HttpStatusCode.BadRequest
+            });
+        
         _context.Pessoas.Remove(result);
         _context.SaveChanges();
-
-        return "Deletada pessoa de " + id; 
+        
+        return Ok(new {
+            msg = "Deletada pessoa de " + id,
+            status = HttpStatusCode.OK
+        });
     }
 
 }
